Moves Clues css into separate file
